backend/pyproject.toml
[project]
name = "osmosis"
version = "0.1.0"
description = "Language learning through osmosis — chat with a cat, learn a language"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "alembic>=1.14",
    "pydantic-settings>=2.7",
    "httpx>=0.28",
    "jinja2>=3.1",
    "fsrs>=4",
    "python-jose[cryptography]>=3.3",
    "authlib>=1.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
